Move check_commit to an saperate file.
