+++ api/common/utils.py
@@ -15,12 +15,11 @@
 from __future__ import print_function
 
 import six
-import os, sys
+import sys
 import traceback
 import numpy as np
 import json
 import collections
-import subprocess
 import itertools
 
 from common import special_op_list
@@ -29,83 +28,6 @@ if six.PY3:
     from . import special_op_list
 else:
     import special_op_list
-
-
-def str2bool(v):
-    if v.lower() in ('yes', 'true', 't', 'y', '1'):
-        return True
-    elif v.lower() in ('no', 'false', 'f', 'n', '0'):
-        return False
-    else:
-        raise argparse.ArgumentTypeError('Unsupported value encountered.')
-
-
-def run_command(command, shell=True):
-    print("run command: %s" % command)
-    p = subprocess.Popen(
-        command, stdout=subprocess.PIPE, stderr=subprocess.STDOUT, shell=shell)
-
-    exit_code = None
-    stdout = ''
-    while exit_code is None or line:
-        exit_code = p.poll()
-        line = p.stdout.readline().decode('utf-8')
-        stdout += line
-
-    return stdout, exit_code
-
-
-def check_commit():
-    try:
-        import tensorflow as tf
-        tf_version = tf.__version__
-    except Exception:
-        tf_version = None
-
-    try:
-        import torch
-        torch_version = torch.__version__
-    except Exception:
-        torch_version = None
-
-    try:
-        current_dir = os.getcwd()
-        print("-- Current directory: %s" % current_dir)
-
-        dir_of_this_file = os.path.dirname(os.path.abspath(__file__))
-        print("-- Entering %s" % dir_of_this_file)
-        os.chdir(dir_of_this_file)
-        print("-- Current directory: %s" % os.getcwd())
-        benchmark_commit, _ = run_command("git rev-parse HEAD")
-        benchmark_commit = benchmark_commit.replace("\n", "")
-        benchmark_update_time, _ = run_command("git show -s --format=%ad")
-        benchmark_update_time = benchmark_update_time.replace("\n", "")
-        os.chdir(current_dir)
-        print("-- Current directory: %s" % os.getcwd())
-
-        import paddle
-        paddle_version = paddle.version.full_version
-        paddle_commit = paddle.version.commit
-
-        print(
-            "==========================================================================="
-        )
-        print("-- paddle version             : %s" % paddle_version)
-        print("-- paddle commit              : %s" % paddle_commit)
-
-        if tf_version:
-            print("-- tensorflow version         : %s" % tf_version)
-
-        if torch_version:
-            print("-- pytorch version            : %s" % torch_version)
-
-        print("-- benchmark commit           : %s" % benchmark_commit)
-        print("-- benchmark last update time : %s" % benchmark_update_time)
-        print(
-            "==========================================================================="
-        )
-    except Exception:
-        pass
 
 
 def _compare(output1, output2, atol):
