+++ api/dynamic_tests_v2/elementwise.py
@@ -19,8 +19,25 @@ class ElementwiseConfig(APIConfig):
     def __init__(self, op_type="elementwise"):
         super(ElementwiseConfig, self).__init__(op_type)
         self.api_name = 'add'
-        self.api_list = {'add': 'add'}
+        self.api_list = {
+            'add': 'add',
+            'maximum': 'maximum',
+            'minimum': 'minimum',
+            'multiply': 'multiply',
+            'pow': 'pow',
+            'subtract': 'subtract'
+        }
         self.feed_spec = [{"range": [-1, 1]}, {"range": [-1, 1]}]
+
+    def disabled(self):
+        if self.api_name in ["pow", "maximum", "minimum", "divide"
+                             ] and self.x_dtype == "float16":
+            print(
+                "Warning:\n"
+                "  1. This config is disabled because float16 is not supported for %s.\n"
+                % (self.api_name))
+            return True
+        return super(ElementwiseConfig, self).disabled()
 
     def init_from_json(self, filename, config_id=0, unknown_dim=16):
         super(ElementwiseConfig, self).init_from_json(filename, config_id,
